+++ mcp-server/src/tools/router.rs
@@ -5,15 +5,18 @@ use serde_json::{Value, json};
 
 use crate::{
     catalog::{ExtensionCatalog, extension_summary},
-    helm::{self, HelmChartRef, HelmInstallPlan},
-    k8s::{HelmReleaseDiscovery, KubernetesClient, ResourceListParams},
-    vault::{SecretObject, VaultClient, VaultError, VaultPath, WriteMode},
+    k8s::{KubernetesClient, ResourceListParams},
+    vault::WriteMode,
 };
 
 use super::{
-    ToolDefinition, common::kube_error, common::pod_exec_error, common::success,
-    common::tool_error, common::vault_error, hydra, k8s_summaries, supernode, vault, workloads,
-    workloads::install, workloads::logs, workloads::outputs, workloads::registry,
+    ToolDefinition,
+    args::{optional_string, optional_u32, required_string},
+    common::kube_error,
+    common::success,
+    common::tool_error,
+    hydra, k8s_summaries, supernode, vault, workloads,
+    workloads::{install, logs},
 };
 
 #[derive(Debug, Clone)]
@@ -78,74 +81,20 @@ impl ToolRouter {
             "cluster.events.list" => events_list(arguments).await,
             "extensions.catalog.list" => catalog_list(catalog),
             "extensions.catalog.get" => catalog_get(arguments, catalog),
-            "vault.runtime.metadata.get" => vault_runtime_metadata_get(arguments).await,
-            "vault.runtime.write" => vault_runtime_write(arguments, WriteMode::Replace).await,
-            "vault.runtime.patch" => vault_runtime_write(arguments, WriteMode::Patch).await,
+            "vault.runtime.metadata.get" => vault::runtime_metadata_get(arguments).await,
+            "vault.runtime.write" => vault::runtime_write(arguments, WriteMode::Replace).await,
+            "vault.runtime.patch" => vault::runtime_write(arguments, WriteMode::Patch).await,
             "hydra.keys.generate" => hydra::generate_keys(arguments).await,
-            "workloads.list" => workloads_list(arguments, catalog).await,
-            "workloads.get" => workloads_get(arguments, catalog).await,
+            "workloads.list" => workloads::list::list(arguments, catalog).await,
+            "workloads.get" => workloads::get::get(arguments, catalog).await,
             "workloads.logs.get" => logs::get(arguments).await,
-            "workloads.metrics.get" => workloads_metrics_get(arguments, catalog).await,
-            "workloads.install" => workloads_install(arguments, catalog).await,
+            "workloads.metrics.get" => workloads::metrics::get(arguments, catalog).await,
+            "workloads.install" => install::install(arguments, catalog).await,
             "workloads.upgrade" => workloads::upgrade::upgrade(arguments, catalog).await,
             "workloads.delete" => workloads::delete::delete(arguments, catalog).await,
             "dolos.snapshot.refresh" => workloads::dolos::snapshot_refresh(arguments).await,
             _ => self.not_implemented_result(definition),
         }
-    }
-}
-
-async fn vault_runtime_metadata_get(arguments: Option<&JsonObject>) -> CallToolResult {
-    let path = match runtime_vault_path(arguments) {
-        Ok(path) => path,
-        Err(error) => return error,
-    };
-    let client = match VaultClient::from_env() {
-        Ok(client) => client,
-        Err(error) => return vault_error("vault.runtime.metadata.get", error),
-    };
-
-    match client.runtime_metadata(&path).await {
-        Ok(metadata) => success(json!({
-            "path": metadata.path,
-            "exists": metadata.exists,
-            "keyNames": metadata.key_names,
-            "keyNamesAvailable": metadata.key_names_available,
-            "currentVersion": metadata.current_version,
-        })),
-        Err(error) => vault_error("vault.runtime.metadata.get", error),
-    }
-}
-
-async fn vault_runtime_write(
-    arguments: Option<&JsonObject>,
-    default_mode: WriteMode,
-) -> CallToolResult {
-    let path = match runtime_vault_path(arguments) {
-        Ok(path) => path,
-        Err(error) => return error,
-    };
-    let secret = match secret_argument(arguments) {
-        Ok(secret) => secret,
-        Err(error) => return error,
-    };
-    let mode = match write_mode(arguments, default_mode) {
-        Ok(mode) => mode,
-        Err(error) => return vault_error("vault.runtime.write", error),
-    };
-    let client = match VaultClient::from_env() {
-        Ok(client) => client,
-        Err(error) => return vault_error("vault.runtime.write", error),
-    };
-
-    match client.write_runtime_secret(&path, &secret, mode).await {
-        Ok(receipt) => success(json!({
-            "path": receipt.path,
-            "writtenKeys": receipt.written_keys,
-            "version": receipt.version,
-            "secretValuesReturned": false,
-        })),
-        Err(error) => vault_error("vault.runtime.write", error),
     }
 }
 
@@ -234,451 +183,6 @@ fn catalog_get(arguments: Option<&JsonObject>, catalog: &ExtensionCatalog) -> Ca
     }
 }
 
-async fn workloads_install(
-    arguments: Option<&JsonObject>,
-    catalog: &ExtensionCatalog,
-) -> CallToolResult {
-    let extension_id = match required_string(arguments, "extensionId") {
-        Ok(value) => value,
-        Err(error) => return error,
-    };
-    let release_name = match required_string(arguments, "releaseName") {
-        Ok(value) => value,
-        Err(error) => return error,
-    };
-    let namespace = match required_string(arguments, "namespace") {
-        Ok(value) => value,
-        Err(error) => return error,
-    };
-    let dry_run = optional_bool(arguments, "dryRun").unwrap_or(true);
-
-    let configuration = match required_object(arguments, "configuration") {
-        Ok(value) => value,
-        Err(error) => return error,
-    };
-    let extension = match catalog.get(&extension_id) {
-        Some(extension) => extension,
-        None => {
-            return tool_error(
-                "unknown_extension",
-                format!("extension not found: {extension_id}"),
-                json!({ "extensionId": extension_id }),
-            );
-        }
-    };
-
-    if let Err(error) = validate_configuration_schema(&configuration, &extension.configuration) {
-        return error;
-    }
-
-    if !matches!(
-        extension.id.as_str(),
-        workloads::registry::DOLOS_EXTENSION_ID
-            | workloads::registry::CARDANO_RELAY_EXTENSION_ID
-            | workloads::registry::CARDANO_BLOCK_PRODUCER_EXTENSION_ID
-            | workloads::registry::APEX_FUSION_RELAY_EXTENSION_ID
-            | workloads::registry::APEX_FUSION_BLOCK_PRODUCER_EXTENSION_ID
-            | workloads::registry::HYDRA_NODE_EXTENSION_ID
-    ) && configuration.get("namespace").and_then(Value::as_str) != Some(namespace.as_str())
-    {
-        return tool_error(
-            "invalid_arguments",
-            "configuration.namespace must match namespace",
-            json!({ "namespace": namespace, "configurationNamespace": configuration.get("namespace") }),
-        );
-    }
-
-    let resolved_configuration = install::apply_defaults(extension, Value::Object(configuration));
-    let resolution = match install::resolve_configuration(
-        extension,
-        &namespace,
-        resolved_configuration,
-        dry_run,
-    )
-    .await
-    {
-        Ok(resolution) => resolution,
-        Err(error) => return error,
-    };
-    let helm_values =
-        install::planned_helm_values(extension, &release_name, &resolution.configuration);
-    let chart = HelmChartRef {
-        chart: extension.chart.clone(),
-        version: extension.default_version.clone(),
-    };
-
-    if dry_run {
-        return success(json!({
-            "action": "install",
-            "dryRun": true,
-            "wouldMutate": false,
-            "release": {
-                "name": release_name,
-                "namespace": namespace,
-            },
-            "extension": {
-                "id": extension.id,
-                "name": extension.name,
-                "version": extension.default_version,
-            },
-            "chart": chart,
-            "resolvedConfiguration": resolution.configuration,
-            "helmValues": helm_values,
-            "availableStorageClasses": resolution.available_storage_classes,
-            "recommendedStorageClasses": resolution.recommended_storage_classes,
-            "notes": [
-                "dry-run planning only; no Kubernetes or Helm mutation was performed",
-                "raw Helm values are rejected by the extension configuration schema"
-            ],
-        }));
-    }
-
-    let plan = HelmInstallPlan {
-        release_name: release_name.clone(),
-        namespace: namespace.clone(),
-        chart: chart.clone(),
-        values: helm_values.clone(),
-    };
-    let helm_result = match helm::install(&plan).await {
-        Ok(result) => result,
-        Err(error) => {
-            let helm_details = match &error {
-                helm::HelmInstallError::Failed {
-                    status,
-                    stdout,
-                    stderr,
-                } => json!({
-                    "tool": "workloads.install",
-                    "extensionId": extension.id,
-                    "releaseName": release_name,
-                    "namespace": namespace,
-                    "status": status,
-                    "stdout": stdout,
-                    "stderr": stderr,
-                }),
-                _ => json!({
-                    "tool": "workloads.install",
-                    "extensionId": extension.id,
-                    "releaseName": release_name,
-                    "namespace": namespace,
-                }),
-            };
-            return tool_error("helm_install_failed", error.to_string(), helm_details);
-        }
-    };
-
-    success(json!({
-        "action": "install",
-        "dryRun": false,
-        "wouldMutate": true,
-        "release": {
-            "name": release_name,
-            "namespace": namespace,
-        },
-        "extension": {
-            "id": extension.id,
-            "name": extension.name,
-            "version": extension.default_version,
-        },
-        "chart": chart,
-        "resolvedConfiguration": resolution.configuration,
-        "helmValues": helm_values,
-        "availableStorageClasses": resolution.available_storage_classes,
-        "recommendedStorageClasses": resolution.recommended_storage_classes,
-        "helm": helm_result,
-        "notes": [
-            "Helm upgrade --install completed successfully",
-            "raw Helm values are rejected by the extension configuration schema"
-        ],
-    }))
-}
-
-async fn workloads_list(
-    arguments: Option<&JsonObject>,
-    catalog: &ExtensionCatalog,
-) -> CallToolResult {
-    let namespace = optional_string(arguments, "namespace");
-    let include_control_plane = optional_bool(arguments, "includeControlPlane").unwrap_or(false);
-    let client = match KubernetesClient::try_default().await {
-        Ok(client) => client,
-        Err(error) => return kube_error("workloads.list", error),
-    };
-    let helm_releases = match HelmReleaseDiscovery::new(client.clone())
-        .list_latest(namespace.as_deref(), include_control_plane)
-        .await
-    {
-        Ok(releases) => releases,
-        Err(error) => {
-            return tool_error(
-                "helm_release_discovery_error",
-                error.to_string(),
-                json!({ "tool": "workloads.list" }),
-            );
-        }
-    };
-
-    let workloads = helm_releases
-        .iter()
-        .map(|release| workload_summary(release, catalog))
-        .collect::<Vec<_>>();
-
-    success(json!({
-        "namespace": namespace,
-        "source": "kubernetes-api+helm-secrets",
-        "workloads": workloads,
-    }))
-}
-
-fn workload_summary(release: &crate::k8s::HelmReleaseSummary, catalog: &ExtensionCatalog) -> Value {
-    let catalog_extension =
-        registry::extension_for_release(release, catalog).map(extension_summary);
-
-    json!({
-        "namespace": release.namespace,
-        "name": release.name,
-        "status": release.status,
-        "revision": release.revision,
-        "chart": release.chart,
-        "appVersion": release.app_version,
-        "updated": release.updated,
-        "catalogExtension": catalog_extension,
-    })
-}
-
-async fn workloads_get(
-    arguments: Option<&JsonObject>,
-    catalog: &ExtensionCatalog,
-) -> CallToolResult {
-    let namespace = match required_string(arguments, "namespace") {
-        Ok(value) => value,
-        Err(error) => return error,
-    };
-    let name = match required_string(arguments, "name") {
-        Ok(value) => value,
-        Err(error) => return error,
-    };
-    let client = match KubernetesClient::try_default().await {
-        Ok(client) => client,
-        Err(error) => return kube_error("workloads.get", error),
-    };
-    let helm_release = match HelmReleaseDiscovery::new(client.clone())
-        .get_latest(&namespace, &name)
-        .await
-    {
-        Ok(release) => release,
-        Err(error) => {
-            return tool_error(
-                "helm_release_discovery_error",
-                error.to_string(),
-                json!({ "tool": "workloads.get", "namespace": namespace, "name": name }),
-            );
-        }
-    };
-
-    let deployment = match get_optional(client.get_deployment(&namespace, &name).await) {
-        Ok(value) => value.map(|deployment| k8s_summaries::deployment_summary(&deployment)),
-        Err(error) => return kube_error("workloads.get", error),
-    };
-    let stateful_set = match get_optional(client.get_stateful_set(&namespace, &name).await) {
-        Ok(value) => value.map(|stateful_set| k8s_summaries::stateful_set_summary(&stateful_set)),
-        Err(error) => return kube_error("workloads.get", error),
-    };
-    let pod = match get_optional(client.get_pod(&namespace, &name).await) {
-        Ok(value) => value.map(|pod| k8s_summaries::pod_summary(&pod)),
-        Err(error) => return kube_error("workloads.get", error),
-    };
-    let services = match client
-        .list_services(
-            Some(&namespace),
-            &ResourceListParams {
-                label_selector: Some(format!("app.kubernetes.io/instance={name}")),
-                ..Default::default()
-            },
-        )
-        .await
-    {
-        Ok(items) => items,
-        Err(error) => return kube_error("workloads.get", error),
-    };
-    let pod_items = match client
-        .list_pods(
-            Some(&namespace),
-            &ResourceListParams {
-                label_selector: Some(format!("app.kubernetes.io/instance={name}")),
-                ..Default::default()
-            },
-        )
-        .await
-    {
-        Ok(items) => items.items,
-        Err(error) => return kube_error("workloads.get", error),
-    };
-    let pods = pod_items
-        .iter()
-        .map(k8s_summaries::pod_summary)
-        .collect::<Vec<_>>();
-
-    if deployment.is_none()
-        && stateful_set.is_none()
-        && pod.is_none()
-        && helm_release.is_none()
-        && pods.is_empty()
-        && services.items.is_empty()
-    {
-        return tool_error(
-            "not_found",
-            format!("workload not found: {namespace}/{name}"),
-            json!({ "namespace": namespace, "name": name }),
-        );
-    }
-
-    success(json!({
-        "namespace": namespace,
-        "name": name,
-        "source": "kubernetes-api+helm-secrets",
-        "helmRelease": helm_release,
-        "deployment": deployment,
-        "statefulSet": stateful_set,
-        "pod": pod,
-        "relatedPods": pods,
-        "logTargets": k8s_summaries::pod_log_target_summaries(&pod_items),
-        "relatedServices": k8s_summaries::service_summaries(services.clone(), true),
-        "outputs": outputs::outputs_for_release(
-            &namespace,
-            &name,
-            helm_release.as_ref(),
-            &services.items,
-            catalog,
-        ),
-    }))
-}
-
-async fn workloads_metrics_get(
-    arguments: Option<&JsonObject>,
-    catalog: &ExtensionCatalog,
-) -> CallToolResult {
-    let namespace = match required_string(arguments, "namespace") {
-        Ok(value) => value,
-        Err(error) => return error,
-    };
-    let workload = match required_string(arguments, "workload") {
-        Ok(value) => value,
-        Err(error) => return error,
-    };
-    let client = match KubernetesClient::try_default().await {
-        Ok(client) => client,
-        Err(error) => return kube_error("workloads.metrics.get", error),
-    };
-    let helm_release = match HelmReleaseDiscovery::new(client.clone())
-        .get_latest(&namespace, &workload)
-        .await
-    {
-        Ok(Some(release)) => release,
-        Ok(None) => {
-            return tool_error(
-                "not_found",
-                format!("workload not found: {namespace}/{workload}"),
-                json!({ "namespace": namespace, "workload": workload }),
-            );
-        }
-        Err(error) => {
-            return tool_error(
-                "helm_release_discovery_error",
-                error.to_string(),
-                json!({ "tool": "workloads.metrics.get", "namespace": namespace, "workload": workload }),
-            );
-        }
-    };
-    let extension = match registry::extension_for_release(&helm_release, catalog) {
-        Some(extension) => extension,
-        None => {
-            return tool_error(
-                "unsupported_metrics_workload",
-                "metrics are only supported for catalog-managed workloads",
-                json!({
-                    "namespace": namespace,
-                    "workload": workload,
-                    "chart": helm_release.chart,
-                }),
-            );
-        }
-    };
-    let metrics_collection = match extension.metrics_collection.as_ref() {
-        Some(metrics_collection) => metrics_collection,
-        None => {
-            return tool_error(
-                "unsupported_metrics_workload",
-                "this extension does not define metrics collection metadata",
-                json!({
-                    "namespace": namespace,
-                    "workload": workload,
-                    "extensionId": extension.id,
-                    "chart": helm_release.chart,
-                }),
-            );
-        }
-    };
-    let pod_name = match find_workload_pod(&client, &namespace, &workload).await {
-        Ok(Some(pod_name)) => pod_name,
-        Ok(None) => {
-            return tool_error(
-                "not_found",
-                format!("no pod found for workload: {namespace}/{workload}"),
-                json!({ "namespace": namespace, "workload": workload }),
-            );
-        }
-        Err(error) => return kube_error("workloads.metrics.get", error),
-    };
-    let command = metrics_collection
-        .command
-        .iter()
-        .map(String::as_str)
-        .collect::<Vec<_>>();
-    let output = match client
-        .pod_exec_capture(
-            &namespace,
-            &pod_name,
-            &metrics_collection.container,
-            &command,
-        )
-        .await
-    {
-        Ok(output) => output,
-        Err(error) => return pod_exec_error("workloads.metrics.get", error),
-    };
-    let metrics = match serde_json::from_str::<Value>(output.stdout.trim()) {
-        Ok(metrics) => metrics,
-        Err(error) => {
-            return tool_error(
-                "invalid_metrics_payload",
-                format!("metrics script did not return valid JSON: {error}"),
-                json!({
-                    "namespace": namespace,
-                    "workload": workload,
-                    "pod": pod_name,
-                    "container": metrics_collection.container,
-                }),
-            );
-        }
-    };
-
-    success(json!({
-        "namespace": namespace,
-        "workload": workload,
-        "pod": pod_name,
-        "container": metrics_collection.container,
-        "source": format!("pod-exec:{}", metrics_collection.command.join(" ")),
-        "extension": {
-            "id": extension.id,
-            "name": extension.name,
-            "version": extension.default_version,
-        },
-        "helmRelease": helm_release,
-        "metrics": metrics,
-        "stderr": if output.stderr.trim().is_empty() { Value::Null } else { Value::String(output.stderr) },
-    }))
-}
-
 fn tool_from_definition(definition: ToolDefinition) -> Tool {
     Tool::new(
         definition.name,
@@ -722,347 +226,12 @@ fn tool_meta(definition: ToolDefinition) -> Meta {
     Meta(meta)
 }
 
-async fn find_workload_pod(
-    client: &KubernetesClient,
-    namespace: &str,
-    workload: &str,
-) -> Result<Option<String>, kube::Error> {
-    let mut pods = client
-        .list_pods(
-            Some(namespace),
-            &ResourceListParams {
-                label_selector: Some(format!("app.kubernetes.io/instance={workload}")),
-                ..Default::default()
-            },
-        )
-        .await?
-        .items;
-
-    if pods.is_empty()
-        && let Some(pod) = get_optional(client.get_pod(namespace, workload).await)?
-    {
-        pods.push(pod);
-    }
-
-    pods.sort_by(|left, right| left.metadata.name.cmp(&right.metadata.name));
-
-    let running_pod = pods
-        .iter()
-        .find(|pod| {
-            pod.status
-                .as_ref()
-                .and_then(|status| status.phase.as_deref())
-                == Some("Running")
-        })
-        .and_then(|pod| pod.metadata.name.clone());
-
-    if running_pod.is_some() {
-        return Ok(running_pod);
-    }
-
-    Ok(pods
-        .iter()
-        .find(|pod| pod.metadata.deletion_timestamp.is_none())
-        .and_then(|pod| pod.metadata.name.clone()))
-}
-
-fn get_optional<T>(result: Result<T, kube::Error>) -> Result<Option<T>, kube::Error> {
-    match result {
-        Ok(value) => Ok(Some(value)),
-        Err(kube::Error::Api(error)) if error.code == 404 => Ok(None),
-        Err(error) => Err(error),
-    }
-}
-
 fn list_params(arguments: Option<&JsonObject>, default_limit: Option<u32>) -> ResourceListParams {
     ResourceListParams {
         label_selector: optional_string(arguments, "labelSelector"),
         field_selector: optional_string(arguments, "fieldSelector"),
         limit: optional_u32(arguments, "limit").or(default_limit),
     }
-}
-
-fn required_object(
-    arguments: Option<&JsonObject>,
-    name: &str,
-) -> Result<JsonObject, CallToolResult> {
-    match arguments.and_then(|arguments| arguments.get(name)) {
-        Some(Value::Object(value)) => Ok(value.clone()),
-        Some(value) => Err(tool_error(
-            "invalid_arguments",
-            format!("expected object argument: {name}"),
-            json!({ "argument": name, "actualType": value_type_name(value) }),
-        )),
-        None => Err(tool_error(
-            "invalid_arguments",
-            format!("missing required object argument: {name}"),
-            json!({ "argument": name }),
-        )),
-    }
-}
-
-fn required_string(arguments: Option<&JsonObject>, name: &str) -> Result<String, CallToolResult> {
-    optional_string(arguments, name).ok_or_else(|| {
-        tool_error(
-            "invalid_arguments",
-            format!("missing required string argument: {name}"),
-            json!({ "argument": name }),
-        )
-    })
-}
-
-fn validate_configuration_schema(
-    values: &JsonObject,
-    schema: &Value,
-) -> Result<(), CallToolResult> {
-    validate_object_value("", values, schema, schema)
-}
-
-fn validate_object_value(
-    path: &str,
-    values: &JsonObject,
-    schema: &Value,
-    root_schema: &Value,
-) -> Result<(), CallToolResult> {
-    let schema = dereference_schema(schema, root_schema);
-    let schema = schema.as_object().ok_or_else(|| {
-        tool_error(
-            "catalog_schema_error",
-            "extension configuration schema must be an object schema",
-            json!({}),
-        )
-    })?;
-    let properties = schema.get("properties").and_then(Value::as_object);
-
-    let additional_properties = schema.get("additionalProperties");
-    if additional_properties.and_then(Value::as_bool) == Some(false) {
-        for key in values.keys() {
-            if !properties.is_some_and(|properties| properties.contains_key(key)) {
-                let field = field_path(path, key);
-                return Err(tool_error(
-                    "invalid_extension_configuration",
-                    format!("unknown extension configuration value: {field}"),
-                    json!({ "field": field }),
-                ));
-            }
-        }
-    }
-
-    if let Some(required) = schema.get("required").and_then(Value::as_array) {
-        for field in required.iter().filter_map(Value::as_str) {
-            if !values.contains_key(field) {
-                let field = field_path(path, field);
-                return Err(tool_error(
-                    "invalid_extension_configuration",
-                    format!("missing required extension configuration value: {field}"),
-                    json!({ "field": field }),
-                ));
-            }
-        }
-    }
-
-    for (key, value) in values {
-        let field = field_path(path, key);
-        if let Some(property_schema) = properties.and_then(|properties| properties.get(key)) {
-            validate_property_value(&field, value, property_schema, root_schema)?;
-        } else if let Some(additional_schema) = additional_properties.and_then(Value::as_object) {
-            validate_property_value(
-                &field,
-                value,
-                &Value::Object(additional_schema.clone()),
-                root_schema,
-            )?;
-        }
-    }
-
-    Ok(())
-}
-
-fn validate_property_value(
-    name: &str,
-    value: &Value,
-    schema: &Value,
-    root_schema: &Value,
-) -> Result<(), CallToolResult> {
-    let schema = dereference_schema(schema, root_schema);
-    if let Some(expected_type) = schema.get("type") {
-        let matches = expected_type_matches(value, expected_type);
-
-        if !matches {
-            return Err(tool_error(
-                "invalid_extension_configuration",
-                format!("invalid type for extension configuration value: {name}"),
-                json!({
-                    "field": name,
-                    "expectedType": expected_type,
-                    "actualType": value_type_name(value),
-                }),
-            ));
-        }
-    }
-
-    if let Some(allowed_values) = schema.get("enum").and_then(Value::as_array)
-        && !allowed_values.iter().any(|allowed| allowed == value)
-    {
-        return Err(tool_error(
-            "invalid_extension_configuration",
-            format!("unsupported value for extension configuration field: {name}"),
-            json!({
-                "field": name,
-                "allowedValues": allowed_values,
-                "actualValue": value,
-            }),
-        ));
-    }
-
-    if let (Some(value), Some(min_length)) = (
-        value.as_str(),
-        schema.get("minLength").and_then(Value::as_u64),
-    ) && value.chars().count() < min_length as usize
-    {
-        return Err(tool_error(
-            "invalid_extension_configuration",
-            format!("extension configuration value is too short: {name}"),
-            json!({
-                "field": name,
-                "minLength": min_length,
-            }),
-        ));
-    }
-
-    if let Some(values) = value.as_object() {
-        validate_object_value(name, values, schema, root_schema)?;
-    } else if let (Some(items), Some(item_schema)) = (value.as_array(), schema.get("items")) {
-        for (index, item) in items.iter().enumerate() {
-            validate_property_value(&format!("{name}[{index}]"), item, item_schema, root_schema)?;
-        }
-    }
-
-    Ok(())
-}
-
-fn dereference_schema<'a>(schema: &'a Value, root_schema: &'a Value) -> &'a Value {
-    let Some(reference) = schema.get("$ref").and_then(Value::as_str) else {
-        return schema;
-    };
-    let Some(name) = reference
-        .strip_prefix("#/definitions/")
-        .or_else(|| reference.strip_prefix("#/$defs/"))
-    else {
-        return schema;
-    };
-
-    root_schema
-        .get("definitions")
-        .or_else(|| root_schema.get("$defs"))
-        .and_then(|definitions| definitions.get(name))
-        .unwrap_or(schema)
-}
-
-fn expected_type_matches(value: &Value, expected_type: &Value) -> bool {
-    if let Some(expected_type) = expected_type.as_str() {
-        return single_type_matches(value, expected_type);
-    }
-
-    expected_type.as_array().is_none_or(|types| {
-        types
-            .iter()
-            .filter_map(Value::as_str)
-            .any(|ty| single_type_matches(value, ty))
-    })
-}
-
-fn single_type_matches(value: &Value, expected_type: &str) -> bool {
-    match expected_type {
-        "array" => value.is_array(),
-        "boolean" => value.is_boolean(),
-        "integer" => value.as_i64().is_some(),
-        "null" => value.is_null(),
-        "number" => value.as_f64().is_some(),
-        "object" => value.is_object(),
-        "string" => value.is_string(),
-        _ => true,
-    }
-}
-
-fn field_path(parent: &str, child: &str) -> String {
-    if parent.is_empty() {
-        child.to_string()
-    } else {
-        format!("{parent}.{child}")
-    }
-}
-
-fn value_type_name(value: &Value) -> &'static str {
-    match value {
-        Value::Null => "null",
-        Value::Bool(_) => "boolean",
-        Value::Number(_) => "number",
-        Value::String(_) => "string",
-        Value::Array(_) => "array",
-        Value::Object(_) => "object",
-    }
-}
-
-fn runtime_vault_path(arguments: Option<&JsonObject>) -> Result<VaultPath, CallToolResult> {
-    let path = required_string(arguments, "path")?;
-    VaultPath::runtime(&path).map_err(|error| {
-        tool_error(
-            "vault_path_not_allowed",
-            error.to_string(),
-            json!({ "path": path }),
-        )
-    })
-}
-
-fn secret_argument(arguments: Option<&JsonObject>) -> Result<SecretObject, CallToolResult> {
-    let value = arguments
-        .and_then(|arguments| arguments.get("values").or_else(|| arguments.get("data")))
-        .cloned()
-        .ok_or_else(|| {
-            tool_error(
-                "invalid_arguments",
-                "missing required secret object argument: values",
-                json!({ "argument": "values" }),
-            )
-        })?;
-
-    SecretObject::new(value).map_err(|error| {
-        tool_error(
-            "invalid_secret_values",
-            error.to_string(),
-            json!({ "argument": "values" }),
-        )
-    })
-}
-
-fn write_mode(
-    arguments: Option<&JsonObject>,
-    default_mode: WriteMode,
-) -> Result<WriteMode, VaultError> {
-    match optional_string(arguments, "mode") {
-        Some(mode) => WriteMode::parse(Some(&mode)),
-        None => Ok(default_mode),
-    }
-}
-
-fn optional_string(arguments: Option<&JsonObject>, name: &str) -> Option<String> {
-    arguments?
-        .get(name)?
-        .as_str()
-        .filter(|value| !value.trim().is_empty())
-        .map(str::to_string)
-}
-
-fn optional_bool(arguments: Option<&JsonObject>, name: &str) -> Option<bool> {
-    arguments?.get(name)?.as_bool()
-}
-
-fn optional_u32(arguments: Option<&JsonObject>, name: &str) -> Option<u32> {
-    arguments?
-        .get(name)?
-        .as_u64()
-        .and_then(|value| u32::try_from(value).ok())
 }
 
 impl Default for ToolRouter {
@@ -1244,7 +413,7 @@ mod tests {
             config: None,
         };
 
-        let summary = workload_summary(&release, &catalog);
+        let summary = workloads::list::workload_summary(&release, &catalog);
 
         assert_eq!(summary.pointer("/name"), Some(&json!("dolos-preview")));
         assert_eq!(
