+++ mcp-server/src/tools/workloads/mod.rs
@@ -6,8 +6,11 @@ use std::collections::BTreeSet;
 
 pub(crate) mod delete;
 pub(crate) mod dolos;
+pub(crate) mod get;
 pub(crate) mod install;
+pub(crate) mod list;
 pub(crate) mod logs;
+pub(crate) mod metrics;
 pub(crate) mod outputs;
 pub(crate) mod registry;
 pub(crate) mod upgrade;
@@ -32,7 +35,7 @@ pub fn definitions() -> &'static [ToolDefinition] {
             approval_class: ApprovalClass::Discovery,
             read_only: true,
             destructive: false,
-            input_schema: r#"{"type":"object","required":["namespace","name"],"properties":{"namespace":{"type":"string"},"name":{"type":"string"}},"additionalProperties":false}"#,
+            input_schema: r#"{"type":"object","required":["namespace","name"],"properties":{"namespace":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63},"name":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63}},"additionalProperties":false}"#,
         },
         ToolDefinition {
             name: "workloads.logs.get",
@@ -42,7 +45,7 @@ pub fn definitions() -> &'static [ToolDefinition] {
             approval_class: ApprovalClass::ReadOnlyDebug,
             read_only: true,
             destructive: false,
-            input_schema: r#"{"type":"object","required":["namespace","workload"],"properties":{"namespace":{"type":"string"},"workload":{"type":"string"},"pod":{"type":"string","description":"Exact pod name to read. Required when the workload has multiple active pods."},"container":{"type":"string","description":"Exact container or init-container name to read. Required when the selected pod has multiple loggable containers."},"tailLines":{"type":"integer","minimum":1,"maximum":1000},"sinceSeconds":{"type":"integer","minimum":1},"previous":{"type":"boolean"},"timestamps":{"type":"boolean"}},"additionalProperties":false}"#,
+            input_schema: r#"{"type":"object","required":["namespace","workload"],"properties":{"namespace":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63},"workload":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63},"pod":{"type":"string","description":"Exact pod name to read. Required when the workload has multiple active pods."},"container":{"type":"string","description":"Exact container or init-container name to read. Required when the selected pod has multiple loggable containers."},"tailLines":{"type":"integer","minimum":1,"maximum":1000},"sinceSeconds":{"type":"integer","minimum":1},"previous":{"type":"boolean"},"timestamps":{"type":"boolean"}},"additionalProperties":false}"#,
         },
         ToolDefinition {
             name: "workloads.metrics.get",
@@ -52,7 +55,7 @@ pub fn definitions() -> &'static [ToolDefinition] {
             approval_class: ApprovalClass::ReadOnlyDebug,
             read_only: true,
             destructive: false,
-            input_schema: r#"{"type":"object","required":["namespace","workload"],"properties":{"namespace":{"type":"string"},"workload":{"type":"string"}},"additionalProperties":false}"#,
+            input_schema: r#"{"type":"object","required":["namespace","workload"],"properties":{"namespace":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63},"workload":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63}},"additionalProperties":false}"#,
         },
         ToolDefinition {
             name: "workloads.install",
@@ -62,7 +65,7 @@ pub fn definitions() -> &'static [ToolDefinition] {
             approval_class: ApprovalClass::Mutation,
             read_only: false,
             destructive: false,
-            input_schema: r#"{"type":"object","required":["extensionId","releaseName","namespace","configuration"],"properties":{"extensionId":{"type":"string","description":"Catalog extension ID to install, for example dolos, cardano-relay, cardano-block-producer, apex-fusion-relay, apex-fusion-block-producer, or hydra-node."},"releaseName":{"type":"string","description":"Helm release name to create or update."},"namespace":{"type":"string","description":"Kubernetes namespace where the workload will be installed."},"configuration":{"type":"object","description":"Required extension-specific configuration object. Use extensions.catalog.get for the selected extensionId and pass values matching that extension configuration schema.","additionalProperties":true},"dryRun":{"type":"boolean","description":"When true, validate and return the install plan without mutating Kubernetes. Defaults to true."}},"additionalProperties":false}"#,
+            input_schema: r#"{"type":"object","required":["extensionId","releaseName","namespace","configuration"],"properties":{"extensionId":{"type":"string","description":"Catalog extension ID to install, for example dolos, cardano-relay, cardano-block-producer, apex-fusion-relay, apex-fusion-block-producer, or hydra-node."},"releaseName":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63,"description":"Helm release name to create or update."},"namespace":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63,"description":"Kubernetes namespace where the workload will be installed."},"configuration":{"type":"object","description":"Required extension-specific configuration object. Use extensions.catalog.get for the selected extensionId and pass values matching that extension configuration schema.","additionalProperties":true},"dryRun":{"type":"boolean","description":"When true, validate and return the install plan without mutating Kubernetes. Defaults to true."}},"additionalProperties":false}"#,
         },
         ToolDefinition {
             name: "workloads.upgrade",
@@ -72,7 +75,7 @@ pub fn definitions() -> &'static [ToolDefinition] {
             approval_class: ApprovalClass::Mutation,
             read_only: false,
             destructive: false,
-            input_schema: r#"{"type":"object","required":["namespace","releaseName","configuration"],"properties":{"namespace":{"type":"string"},"releaseName":{"type":"string"},"configuration":{"type":"object","description":"Required extension-specific configuration object. Use extensions.catalog.get for the installed extension and pass values matching that extension configuration schema.","additionalProperties":true},"dryRun":{"type":"boolean","description":"When true, validate and return the upgrade plan without mutating Kubernetes."}},"additionalProperties":false}"#,
+            input_schema: r#"{"type":"object","required":["namespace","releaseName","configuration"],"properties":{"namespace":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63},"releaseName":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63},"configuration":{"type":"object","description":"Required extension-specific configuration object. Use extensions.catalog.get for the installed extension and pass values matching that extension configuration schema.","additionalProperties":true},"dryRun":{"type":"boolean","description":"When true, validate and return the upgrade plan without mutating Kubernetes."}},"additionalProperties":false}"#,
         },
         ToolDefinition {
             name: "workloads.delete",
@@ -82,7 +85,7 @@ pub fn definitions() -> &'static [ToolDefinition] {
             approval_class: ApprovalClass::Destructive,
             read_only: false,
             destructive: true,
-            input_schema: r#"{"type":"object","required":["namespace","releaseName"],"properties":{"namespace":{"type":"string"},"releaseName":{"type":"string"},"dryRun":{"type":"boolean","description":"When true, validate and return the delete plan without mutating Kubernetes. Defaults to true."},"deletePvcs":{"type":"boolean","description":"Delete candidate PVCs after Helm uninstall. Defaults to false."},"approvalId":{"type":"string"}},"additionalProperties":false}"#,
+            input_schema: r#"{"type":"object","required":["namespace","releaseName"],"properties":{"namespace":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63},"releaseName":{"type":"string","pattern":"^[a-z0-9]([-a-z0-9]*[a-z0-9])?$","maxLength":63},"dryRun":{"type":"boolean","description":"When true, validate and return the delete plan without mutating Kubernetes. Defaults to true."},"deletePvcs":{"type":"boolean","description":"Delete candidate PVCs after Helm uninstall. Defaults to false."},"approvalId":{"type":"string"}},"additionalProperties":false}"#,
         },
     ]
 }
